rename: activationmanager.py
was: modemanager.py
