+++ tests/unit/test_core/test_mode.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock, call
 import pytest
 
 from wakepy.core.mode import Mode, ActivationResult, ModeExit
-from wakepy.core.modemanager import ModeActivationManager
+from wakepy.core.activationmanager import ModeActivationManager
 from wakepy.core.method import Method
